refactor(#794): extract remaining 4 inline test modules post-merge
